linux: move jtag access code into a seperate file
