+++ mrmShared/linux/uio_mrf.c
@@ -3,21 +3,7 @@
  * See file LICENSE for terms.
  */
 
-#include <linux/version.h>
-#include <linux/device.h>
-#include <linux/module.h>
-#include <linux/platform_device.h>
-#include <linux/uio_driver.h>
-#include <linux/io.h>
-#include <linux/interrupt.h>
-#include <linux/pci.h>
-#include <linux/delay.h>
-#ifdef CONFIG_GENERIC_GPIO
-#  include <linux/gpio.h>
-#endif
-#ifdef CONFIG_PARPORT_NOT_PC
-#  include <linux/parport.h>
-#endif
+#include "mrf.h"
 
 #define DRV_NAME "mrf-pci"
 #define DRV_VERSION "1"
@@ -25,119 +11,6 @@
 MODULE_LICENSE("GPL v2");
 MODULE_VERSION(DRV_VERSION);
 MODULE_AUTHOR("Michael Davidsaver <mdavidsaver@bnl.gov>");
-
-#ifdef CONFIG_GENERIC_GPIO
-static int modparam_gpiobase = -1;
-module_param_named(gpiobase, modparam_gpiobase, int, 0440);
-MODULE_PARM_DESC(gpiobase, "The GPIO number base. -1 means dynamic, which is the default.");
-#endif
-
-
-#ifdef CONFIG_PARPORT_NOT_PC
-// parallel port programming cable description
-struct ppcable {
-    const char *name;
-    // bit masks for each JTAG signal
-
-    // output (host -> devices signals)
-    //  may be in data or control byte
-    unsigned char tms_data, tms_ctrl;
-    unsigned char tck_data, tck_ctrl;
-    unsigned char tdi_data, tdi_ctrl;
-
-    // input (device -> host)
-    // may be in status or control byte
-    unsigned char tdo_sts, tdo_ctrl;
-};
-
-static const struct ppcable cables[];
-static char *cablename = "Minimal";
-module_param_named(cable, cablename, charp, 0444);
-MODULE_PARM_DESC(cable, "Name of JTAG parallel port cable to emulate");
-#endif
-
-/************************ Register definitions ****************************/
-
-/*
- * A selection of registers for the PLX PCI9030
- *
- * This device is exposed as BAR #0 on PCI and PMC
- * versions of the EVR
- */
-
-/* Address space #0 is exposed as BAR #2 */
-#define LAS0BRD  0x28
-/* Set for big endian, clear for little endian (swapped) */
-#  define LAS0BRD_ENDIAN 0x01000000
-
-/* Interrupt control */
-#define INTCSR   0x4c
-#  define INTCSR_INT1_Enable   0x01
-#  define INTCSR_INT1_Polarity 0x02
-#  define INTCSR_INT1_Status   0x04
-#  define INTCSR_INT2_Enable   0x08
-#  define INTCSR_INT2_Polarity 0x10
-#  define INTCSR_INT2_Status   0x20
-#  define INTCSR_PCI_Enable    0x40
-#  define INTCSR_SW_INTR       0x80
-
-#define GPIOC   0x54
-#  define GPIOC_pin0_fn   1
-#  define GPIOC_pin0_dir  2
-#  define GPIOC_pin0_data 4
-#  define GPIOC_pin_fn(N)   (GPIOC_pin0_fn<<(3*(N)))
-#  define GPIOC_pin_dir(N)  (GPIOC_pin0_dir<<(3*(N)))
-#  define GPIOC_pin_data(N) (GPIOC_pin0_data<<(3*(N)))
-/* 8 are supported, but only 4 are used.
- * 0 - TCLK, 1 - TMS, 2 - TDO, 3 - TDI
- */
-#  define GPIOC_num_pins 4
-
-/*
- * A selection of registers for the PLX PCI9056
- *
- * This device is exposed as BAR #0
- */
-
-#define BIGEND9056 0x0C // 8 bit
-#  define BIGEND9056_BIG (1<<2)
-
-#define INTCSR9056 0x68 // 32 bit
-#  define INTCSR9056_PCI_Enable (1<<8)
-
-#  define INTCSR9056_DBL_Enable (1<<9)
-#  define INTCSR9056_ABT_Enable (1<<10)
-#  define INTCSR9056_LCL_Enable (1<<11)
-#  define INTCSR9056_DBL_Status (1<<13) /* PCI doorbell */
-#  define INTCSR9056_ABT_Status (1<<14) /* PCI abort */
-#  define INTCSR9056_LCL_Status (1<<15) /* local */
-
-#  define INTCSR9056_LBL_Enable (1<<17)
-#  define INTCSR9056_DM0_Enable (1<<18)
-#  define INTCSR9056_DM1_Enable (1<<19)
-#  define INTCSR9056_LBL_Status (1<<20) /* Local doorbell */
-#  define INTCSR9056_DM0_Status (1<<21) /* DMA 0 */
-#  define INTCSR9056_DM1_Status (1<<22) /* DMA 1 */
-
-#  define INTCSR9056_Status (INTCSR9056_DBL_Status| \
-    INTCSR9056_ABT_Status|INTCSR9056_LCL_Status|INTCSR9056_LBL_Status| \
-    INTCSR9056_DM0_Status|INTCSR9056_DM1_Status)
-
-/* For MRM EVR 230 and 300 series
- */
-
-#define IRQFlag     0x008
-#  define IRQ_LinkChg   0x40
-#  define IRQ_BufFull   0x20
-#  define IRQ_HWMapped  0x10
-#  define IRQ_Event     0x08
-#  define IRQ_Heartbeat 0x04
-#  define IRQ_FIFOFull  0x02
-#  define IRQ_RXErr     0x01
-
-#define IRQEnable   0x00c
-/* Same bits as IRQFlag plus */
-#  define IRQ_Enable    0x80000000
 
 /************************ Compatability ****************************/
 
@@ -153,356 +26,6 @@ void __iomem *pci_ioremap_bar(struct pci_dev* pdev,int bar)
         return ioremap_nocache(pci_resource_start(pdev,bar),
                 pci_resource_len(pdev,bar));
 }
-#endif
-
-/**************************** PCI Driver *************************/
-
-
-struct mrf_priv {
-    struct uio_info uio;
-#if defined(CONFIG_GENERIC_GPIO) || defined(CONFIG_PARPORT_NOT_PC)
-    spinlock_t lock;
-#endif
-#ifdef CONFIG_GENERIC_GPIO
-    struct gpio_chip gpio;
-    int gpio_cleanup;
-#endif
-#ifdef CONFIG_PARPORT_NOT_PC
-    unsigned int ppenable;
-    const struct ppcable *cable;
-    struct parport *port;
-#endif
-    struct pci_dev *pdev;
-};
-
-#ifdef CONFIG_GENERIC_GPIO
-/************************* GPIO for JTAG **************************/
-
-static
-int mrf_gpio_request(struct gpio_chip *chip, unsigned offset)
-{
-    if(offset>=GPIOC_num_pins)
-        return -EINVAL;
-    return 0;
-}
-
-static
-void mrf_gpio_free(struct gpio_chip *chip, unsigned offset)
-{}
-
-static
-int mrf_gpio_get(struct gpio_chip *chip, unsigned offset)
-{
-    int ret;
-    struct mrf_priv *priv = container_of(chip, struct mrf_priv, gpio);
-    void __iomem *plx = priv->uio.mem[0].internal_addr;
-
-    if(offset>=GPIOC_num_pins)
-        return 0;
-
-    ret = ioread32(plx + GPIOC) & GPIOC_pin_data(offset);
-
-    return !!ret;
-}
-
-static
-void mrf_gpio_set(struct gpio_chip *chip, unsigned offset, int value)
-{
-    u32 val;
-    struct mrf_priv *priv = container_of(chip, struct mrf_priv, gpio);
-    void __iomem *plx = priv->uio.mem[0].internal_addr;
-
-    if(offset==2 || offset>=GPIOC_num_pins)
-        return;
-
-    spin_lock(&priv->lock);
-
-    val = ioread32(plx + GPIOC);
-    if(value)
-        val |= GPIOC_pin_data(offset);
-    else
-        val &= ~GPIOC_pin_data(offset);
-
-    iowrite32(val, plx + GPIOC);
-
-    spin_unlock(&priv->lock);
-}
-
-static
-int mrf_gpio_dir_in(struct gpio_chip *chip, unsigned offset)
-{
-    if(offset!=2) /* Only TDO may be input */
-        return -EINVAL;
-    return 0;
-}
-
-static
-int mrf_gpio_dir_out(struct gpio_chip *chip, unsigned offset, int value)
-{
-    if(offset==2) /* All but TDO are output */
-        return -EINVAL;
-    mrf_gpio_set(chip, offset, value);
-    return 0;
-}
-
-static
-void mrf_gpio_setup(struct mrf_priv *priv)
-{
-    int ret;
-    /* Setup GPIO config */
-    priv->gpio.label = dev_name(&priv->pdev->dev);
-    priv->gpio.owner = THIS_MODULE;
-    priv->gpio.base = modparam_gpiobase;
-    priv->gpio.ngpio = GPIOC_num_pins;
-    priv->gpio.can_sleep = 0;
-    priv->gpio.request = &mrf_gpio_request;
-    priv->gpio.free = &mrf_gpio_free;
-    priv->gpio.direction_input = &mrf_gpio_dir_in;
-    priv->gpio.direction_output = &mrf_gpio_dir_out;
-    priv->gpio.set = &mrf_gpio_set;
-    priv->gpio.get = &mrf_gpio_get;
-
-    /* gpiochip register fail is not fatal.  Likely due
-     * to collision of gpiobase numbers.  Must use dynamic
-     * numbers when multiple MRF cards w/ PLX 9030 are present.
-     */
-    ret = gpiochip_add(&priv->gpio);
-    if (ret) {
-        dev_warn(&priv->pdev->dev, "GPIO setup error, JTAG unavailable\n");
-    } else {
-        priv->gpio_cleanup = 1;
-        dev_info(&priv->pdev->dev, "GPIO setup ok, JTAG available at bit %d\n",
-                 priv->gpio.base);
-    }
-}
-
-static
-void mrf_gpio_cleanup(struct mrf_priv *priv)
-{
-    if (priv->gpio_cleanup) {
-        int status = gpiochip_remove(&priv->gpio);
-        while (status==-EBUSY) {
-            msleep(10); /* operation in progress... zzzzz */
-            status = gpiochip_remove(&priv->gpio);
-        }
-        BUG_ON(status && status!=-EBUSY);
-        dev_info(&priv->pdev->dev, "GPIO Cleaned up\n");
-    }
-}
-#endif
-
-/******************** Parallel port for JTAG **********************/
-#ifdef CONFIG_PARPORT_NOT_PC
-
-/* if this signal SIG is included in register REG (non-zero mask)
- * then set/clear the appropriate HW bit.
- * Translation between cable bit mask and GPIO bit mask
- */
-#define PPBITSET(REG, SIG, BIT) \
-  if(dev->cable->SIG ## _ ## REG)  {if(d & dev->cable->SIG ## _ ## REG) \
-    {val |= GPIOC_pin_data(BIT);} else {val &= ~GPIOC_pin_data(BIT);}}
-
-/* if this signal SIG is included in register REG (non-zero mask)
- * and the GPIO bit is active, then set the appropriate cable bit
- */
-#define PPBITGET(REG, SIG, BIT) \
-  if(dev->cable->SIG ## _ ## REG && val & GPIOC_pin_data(BIT)) \
-  {ret |= dev->cable->SIG ## _ ## REG;}
-
-static void mrfpp_write_data(struct parport *port, unsigned char d)
-{
-    struct mrf_priv *dev = port->private_data;
-    void __iomem *plx = dev->uio.mem[0].internal_addr;
-    u32 val;
-            
-    spin_lock(&dev->lock);
-    if(!dev->ppenable) {
-        spin_unlock(&dev->lock);
-        return;
-    }
-    val = ioread32(plx + GPIOC);
-
-    PPBITSET(data, tck, 0)
-    PPBITSET(data, tms, 1)
-    PPBITSET(data, tdi, 3)
-    // Don't write TDO
-
-    iowrite32(val, plx + GPIOC);
-    spin_unlock(&dev->lock);
-}
-
-static void mrfpp_write_control(struct parport *port, unsigned char d)
-{
-    struct mrf_priv *dev = port->private_data;
-    void __iomem *plx = dev->uio.mem[0].internal_addr;
-    u32 val;
-
-    d ^= 0x0b; // bits are hardware inverted
-
-    spin_lock(&dev->lock);
-    if(!dev->ppenable) {
-        spin_unlock(&dev->lock);
-        return;
-    }
-    val = ioread32(plx + GPIOC);
-
-    PPBITSET(ctrl, tck, 0)
-    PPBITSET(ctrl, tms, 1)
-    PPBITSET(ctrl, tdi, 3)
-    // Don't write TDO
-
-    iowrite32(val, plx + GPIOC);
-    spin_unlock(&dev->lock);
-}
-
-static unsigned char mrfpp_read_data(struct parport *port)
-{
-    unsigned char ret = 0;
-    struct mrf_priv *dev = port->private_data;
-    void __iomem *plx = dev->uio.mem[0].internal_addr;
-    u32 val;
-
-    val = ioread32(plx + GPIOC);
-    PPBITGET(data, tck, 0)
-    PPBITGET(data, tms, 1)
-    PPBITGET(data, tdi, 3)
-
-    return ret;
-}
-
-static unsigned char mrfpp_read_control(struct parport *port)
-{
-    unsigned char ret = 0;
-    struct mrf_priv *dev = port->private_data;
-    void __iomem *plx = dev->uio.mem[0].internal_addr;
-    u32 val;
-
-    val = ioread32(plx + GPIOC);
-    PPBITGET(ctrl, tck, 0)
-    PPBITGET(ctrl, tms, 1)
-    PPBITGET(ctrl, tdo, 2)
-    PPBITGET(ctrl, tdi, 3)
-    ret ^= 0x0b; // bits are hardware inverted
-
-    return ret;
-}
-
-static unsigned char mrfpp_read_status(struct parport *port)
-{
-    unsigned char ret = 0;
-    struct mrf_priv *dev = port->private_data;
-    void __iomem *plx = dev->uio.mem[0].internal_addr;
-    u32 val;
-
-    val = ioread32(plx + GPIOC);
-    PPBITGET(sts, tdo, 2)
-    ret ^= 0x80; // BUSY line is hardware inverted
-
-    return ret;
-}
-
-static unsigned char mrfpp_frob_control(struct parport *dev, unsigned char mask,
-                                        unsigned char val)
-{
-    unsigned char ret;
-
-    // TODO, not atomic
-    ret = mrfpp_read_control(dev);
-    ret = (ret & ~mask) ^ val;
-    mrfpp_write_control(dev, ret);
-    return ret;
-}
-
-static void mrfpp_noop(struct parport *dev) {}
-
-static void mrfpp_initstate(struct pardevice *dev, struct parport_state *s)
-{
-    // New clients clear all JTAG bits
-    s->u.pc.ctr = 0x0b;
-    s->u.pc.ecr = 0;
-}
-
-static void mrfpp_savestate(struct parport *dev, struct parport_state *s)
-{
-    s->u.pc.ctr = mrfpp_read_control(dev);
-    s->u.pc.ecr = mrfpp_read_data(dev);
-}
-static void mrfpp_restorestate(struct parport *dev, struct parport_state *s)
-{
-    mrfpp_write_control(dev, s->u.pc.ctr);
-    mrfpp_write_data(dev, s->u.pc.ecr);
-}
-
-static struct parport_operations mrfppops = {
-    .owner = THIS_MODULE,
-
-    .write_data    = mrfpp_write_data,
-    .write_control = mrfpp_write_control,
-
-    .read_data     = mrfpp_read_data,
-    .read_status   = mrfpp_read_status,
-    .read_control  = mrfpp_read_control,
-
-    .frob_control  = mrfpp_frob_control,
-
-    .enable_irq    = mrfpp_noop,
-    .disable_irq   = mrfpp_noop,
-    .data_forward  = mrfpp_noop,
-    .data_reverse  = mrfpp_noop,
-
-    .init_state    = mrfpp_initstate,
-    .save_state    = mrfpp_savestate,
-    .restore_state = mrfpp_restorestate,
-};
-
-static
-void mrf_pp_setup(struct mrf_priv* dev)
-{
-    const struct ppcable *ccable = cables;
-
-    for(; ccable->name; ccable++) {
-        if(strcmp(ccable->name, cablename)==0)
-            break;
-    }
-    if(!ccable) {
-        dev_err(&dev->pdev->dev, "Cable '%s' is not defined\n", cablename);
-        return;
-    }
-
-    dev->port = parport_register_port(0, PARPORT_IRQ_NONE,
-                                      PARPORT_DMA_NONE,
-                                      &mrfppops);
-
-    if(!dev->port) {
-        dev_err(&dev->pdev->dev, "Failed to register parallel port\n");
-        return;
-    }
-
-    dev->cable = ccable;
-    dev->port->modes = PARPORT_MODE_PCSPP; /* support only basic PC port ops */
-    dev->port->private_data = dev;
-    dev->port->dev = &dev->pdev->dev;
-
-    parport_announce_port(dev->port);
-    /* Protection from whatever random drivers have just tried
-     * to print to us...
-     */
-    dev->ppenable = 1;
-
-    dev_info(&dev->pdev->dev, "Emulating cable: %s\n", ccable->name);
-    return;
-}
-
-static
-void mrf_pp_cleanup(struct mrf_priv* priv)
-{
-    if(priv->port)
-        parport_remove_port(priv->port);
-}
-
-#undef PPBITGET
-#undef PPBITSET
-
 #endif
 
 /******************** PCI interrupt handler ***********************/
@@ -791,22 +314,3 @@ static void __exit mrf_exit_module(void)
 }
 module_exit(mrf_exit_module);
 
-
-static const struct ppcable cables[] =
-{
-  { // Urjtag minimal
-    .name = "Minimal",
-    .tms_data = 1 << 1,
-    .tck_data = 1 << 2,
-    .tdi_data = 1 << 3,
-    .tdo_sts  = 1 << 7,
-  },
-  { // Altera ByteBlaster
-    .name = "ByteBlaster",
-    .tck_data = 1 << 0,
-    .tms_data = 1 << 1,
-    .tdi_data = 1 << 6,
-    .tdo_sts  = 1 << 7,
-  },
-  {NULL}
-};
